Refactor Generator component with ReactGenerator
