+++ src/main/webapp/app/springboot/primary/react-generator/index.ts
@@ -1,0 +1,4 @@
+import ReactGeneratorComponent from './ReactGenerator.component';
+import ReactGeneratorVue from './ReactGenerator.vue';
+
+export { ReactGeneratorComponent, ReactGeneratorVue };
